requirements.txt
beautifulsoup4==4.8.1
lxml==4.4.1
PyYAML==5.1.2
soupsieve==1.9.5
urllib3==1.25.7
